src/pocketpaw/ai_ui/builtins/ai_fast_api/source/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.60.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
aiofiles>=24.1.0
tenacity>=9.0.0
aiolimiter>=1.2.1
python-multipart>=0.0.18
g4f>=7.0.0
zendriver>=0.15.0
